OpenSSL/ppc-amigaos: avoid broken stdatomic.h
For whatever reason, when the toolchain is built against newlib,
GCC does not provide stdint.h, and thus does not pre-define the
likes of __INT_LEAST8_TYPE__ or __UINTPTR_TYPE__. Thus, including
stdatomic.h (which expects these to be defined) fails hard.

Drop this if ever the toolchain is fixed.